+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -8,9 +8,9 @@ import java.util.Objects;
 import java.util.UUID;
 import com.techcourse.db.InMemoryUserRepository;
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.coyote.http11.HttpCookie;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http.HttpCookie;
+import org.apache.coyote.http.HttpRequest;
+import org.apache.coyote.http.HttpResponse;
 import org.apache.coyote.session.Session;
 import org.apache.coyote.session.SessionManager;
 import org.slf4j.Logger;
