+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -6,8 +6,8 @@ import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http.HttpRequest;
+import org.apache.coyote.http.HttpResponse;
 
 public class RegisterController extends AbstractController {
