+++ tomcat/src/main/java/com/techcourse/controller/RequestMapping.java
@@ -2,7 +2,7 @@ package com.techcourse.controller;
 
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http.HttpRequest;
 
 public class RequestMapping {
