+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,6 +7,8 @@ import java.net.Socket;
 import com.techcourse.controller.RequestMapping;
 import com.techcourse.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http.HttpRequest;
+import org.apache.coyote.http.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
